// src/app/enums/lawGroupEnum.ts
export enum LawGroupEnum {
    group1Al = 'LegjislacioniFiskal',
    group1En = 'FiscalLegislation',
    law1Name = 'FISCAL_LEGISLATION',
    group2Al = 'StandartetEKontabilitetit',
    group2En = 'AccountingStandards',
    law2Name = 'ACCOUNTING',
    group3Al = 'KodiIPunes',
    group3En = 'LaborCode',
    law3Name = 'WORK_CODE',
    group4Al = 'LegjislacioniDoganor',
    group4En = 'CustomsLegislation',
    law4Name = 'CUSTOM',

}